strings: Clean up strings
This patch is a step towards having strings which are suitable
for translation. Most of the change is due to removal of the PRIxx
format strings and replacement with %llu etc. This helps to make
the strings easier to understand and also since there was a mix
of the two kind of format specicifer, now everything uses a single
type.

There are some exceptions... I've not bothered making the changes in
programs which are likely to be dropped in the future, since we
don't extract strings from them anyway. Also I've not made any
changes in the library, since we want to eliminate all user output
from the library.

Along the way I've found a few incorrect format specifiers (%lld
instead of %llu) fixed a few strings which incorrect spacing,
merged a few duplicate strings (which were almost the same but
are now exactly the same) and removed one or two strings which
shouldn't be translated.

There is still plenty to do in this area, but this sorts out most
of the easy stuff.

+++ gfs2/fsck/util.c
@@ -75,7 +75,8 @@ void warm_fuzzy_stuff(uint64_t block)
 			seconds = tv.tv_sec;
 			if (last_fs_block) {
 				percent = (block * 100) / last_fs_block;
-				log_notice( _("\r%" PRIu64 " percent complete.\r"), percent);
+				log_notice( _("\r%llu percent complete.\r"),
+					   (unsigned long long)percent);
 				fflush(stdout);
 			}
 		}
